docs/requirements.txt
CommonMark==0.5.4
sphinx-markdown-tables=0.0.9